Added TODO annotations all over the place.
They may not be in a completely politically correct language, but in my
defense, I really didn't remember writing stuff that is this ugly. Or
maybe i don't realize how much better I got in the meantime.
Also, this commit includes a couple of very minor fixes that were quicker
to fix than to annotate.